-fix crash with FdPollImplSet related to recent changes -remove a few compiler warnings

+++ repro/FilterStore.cxx
@@ -379,7 +379,7 @@ FilterStore::applyRegex(int conditionNum, const Data& header, const Data& match,
 
    if (rewrite.find("$") != Data::npos)
    {
-      for (int i=1; i<matches.size(); i++)
+      for (int i=1; i<(int)matches.size(); i++)
       {
          Data subExp(matches[i]);
          DebugLog( << "  subExpression[" <<i <<"]="<< subExp );
